requirements.txt
boto3
botocore
Django==3.2.1
django-cors-headers
djangorestframework
django-admin-interface
whitenoise
gunicorn
psycopg2-binary
requests
pyocclient
pillow
earthengine-api
django-leaflet
rasterio
turfpy
shapely
geopandas
python-decouple
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
earthengine-api
psycopg2
sqlalchemy
celery
django-celery-beat
django-celery-results
redis
django-redis